requirements.txt
Pillow>=9.0.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
nltk>=3.8.0
requests>=2.31.0
